Add files via upload
Explore an imaginary realm in "PalaceQuest"! This text-based game has a playable area made via multi-directionally linked nodes that the user traverses between. Watch out for enemies who may try to stop you, and look out for allies to be made as you try to make it to the Palace... and whatever lies in wait there!